pyproject.toml
[project]
name = "rlvr-invoice-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.2.0",
    "datasets>=4.5.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "jupyterlab>=4.5.3",
    "openai>=2.15.0",
    "pandas>=3.0.0",
    "peft>=0.15.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "torch>=2.0.0",
    "transformers>=4.45.0",
    "trl>=0.15.0",
]
